Fix linuxvcs
linuxvcs.pp attempts to parse the /proc/PID/stat files from the start,
but the second field is essentially free text making that untenable
unfortunately.

This change addresses that by parsing the space separated fields from the
end as they consist exclusively of numbers.  Some of the larger numbers
overflow the int64 array values, but not for any that are concerning.